+++ Changelog.md
@@ -7,6 +7,7 @@
 
 - Added forwarding/proxying of method calls and property getters from the Checkout manager to the underlying store
 - Added `is_active` flag to shipping methods
+- Changed checkout store resolution to happen via the Laravel DI instead of `new SpecificStoreClass()`
 
 ## 3.1.0
 ##### 2022-11-07